backend/requirements.txt
# Pursuit Response Platform - Backend Dependencies
# Python 3.11+

# Web Framework
fastapi[all]>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.31
asyncpg==0.29.0
alembic==1.12.1
chromadb>=0.4.22
anyio>=4.5.0
mcp==1.22.0

# AI SDKs
anthropic>=0.25.0
openai>=1.33.0

# Task Queue
celery[redis]==5.3.4
redis==4.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# HTTP Client
httpx>=0.25.2

# Document Processing
python-docx==1.1.0
python-pptx==0.6.23
pypdf>=3.0.0
pdfplumber==0.10.3

# Data Validation
pydantic>=2.5.2
pydantic-settings>=2.1.0
email-validator==2.1.0

# Utilities
structlog==23.3.0
python-dateutil==2.8.2
tenacity>=8.2.3
mem0ai>=0.0.12

psycopg2-binary>=2.9.9

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

